- Server.SaveAsync(); added. you cann now add a server with a name, a server type and the option on/off at start. (server-image, the os-image to initialize isnt added - ubuntu will be used - there is a problem with the hetzner api).

+++ cloud-api-net/Api/ServerImage.cs
@@ -1,13 +1,25 @@
 ﻿namespace lkcode.hetznercloudapi.Api
 {
+    /// <summary>
+    /// 
+    /// </summary>
     public class ServerImage
     {
-        public int Id { get; set; }
-
+        /// <summary>
+        /// 
+        /// </summary>
+        public long Id { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
         public string Type { get; set; }
-
+        /// <summary>
+        /// 
+        /// </summary>
         public string Name { get; set; }
-
+        /// <summary>
+        /// 
+        /// </summary>
         public string Description { get; set; }
     }
 }
